// src/api/yarb/gen/model/board-note.ts
// tslint:disable
/// <reference path="../custom.d.ts" />
/**
 * YARB API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface BoardNote
 */
export interface BoardNote {
    /**
     * 
     * @type {number}
     * @memberof BoardNote
     */
    id: number;
    /**
     * 
     * @type {Date}
     * @memberof BoardNote
     */
    creationDate: Date;
    /**
     * 
     * @type {string}
     * @memberof BoardNote
     */
    content: string;
    /**
     * 
     * @type {number}
     * @memberof BoardNote
     */
    votes: number;
}