// src/api/yarb/gen/model/update-board-note.ts
// tslint:disable
/// <reference path="../custom.d.ts" />
/**
 * YARB API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface UpdateBoardNote
 */
export interface UpdateBoardNote {
    /**
     * 
     * @type {string}
     * @memberof UpdateBoardNote
     */
    content: string;
}